started the port to higher order component by changing the name etc for licensing purposes (as far as I could understand them)
