+++ README.md
@@ -1,6 +1,6 @@
-# ampersand-react-mixin
+# ampersand-react-adapter
 
-Mixin for react classes to easily listen for changes and re-render from ampersand models/collections.
+Higher order component for react classes to easily listen for changes and re-render from ampersand models/collections.
 
 
 ## How it works
@@ -16,13 +16,13 @@ Or you can explicitly call `watch` on the component and pass an ampersand model,
 If it's a collection, it listens to `'add remove reset sort'` events. If it's a `State` object, it listens to `'change'` events.
 
 ## install
-
+[//]: # (TODO ishan 2019-02-20 Update the installation instructions)
 ```
 npm install ampersand-react-mixin
 ```
 
 ## example
-
+[//]: # (TODO ishan 2019-02-20 Update the example usage) 
 ```javascript
 var React = require('react');
 var ampersandMixin = require('ampersand-react-mixin');
@@ -42,3 +42,5 @@ module.exports = React.createClass({
 
 MIT
 
+## NOTE
+This is an ongoing effort to take the original `ampersand-react-mixin` library and port it to React high-order component
